Separate the object vertices and orientation data
This enables us to load the objects from the resources independent of their
orientation. This diff introduces the concept of an ObjectModel - a
representation of the data in the .obj file - and a Sprite - an entity in the
scene. The orientation is maintained in the sprite independent of the
ObjectModel, which can enable the sharing of ObjectModels across different
sprites - e.g just one black/white pawn instead of 8.

Change-Id: Id0007d471f1efbe64db8dbaabae81aabb5c45f41
Reviewed-on: https://r.enimai.com/2383
Reviewed-by: Srinivasan Sridharan <srini@enimai.com>
Tested-by: Rahul Iyer <rni@enimai.com>