OpenSpatial Refactor: De-dupe redundent things
- Instead of having Map<Device, foo>, Map<Device, bar> etc.
all of the callbacks for each device live in one EventCallbacks
object.

- Instead of having one callback per event type have a general
one that handles all events.

- Instead of having n (un)registerFor<n>Events functions there
is now 1 set of (un)registerForEvents that takes event type as
an argument.

- Instead of having a 2 registration intents per event type
(register/unregister) there is now one registration change intent
that takes the event type and whether you are trying to register or
unregister as a parameter.

- Instead of having one intent per event type for data received
there is now one intent with the event types being added as
parameters.

Change-Id: I788c029930104e86a4a8d2552212a7a06950cda6
Reviewed-on: https://r.nod.com/4094
Reviewed-by: Josh Bailey <josh@nod-labs.com>
Tested-by: Josh Bailey <josh@nod-labs.com>