Improve enable/disable cleanup handling
In the event that a device is encountered which is reporting
data that was not asked for the service will now attempt to disable
data reporting for all devices in order to put things in to a known
state. The cleanup methods have also been fixed so that no needless
control commands are issued.

Change-Id: Ifa126cab1153689fa15bcb4bb46d81e71e70ff80
Reviewed-on: https://r.nod.com/4837
Reviewed-by: Peter Williams <peter@nod-labs.com>
Tested-by: Josh Bailey <josh@nod-labs.com>